backend/requirements.txt
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
click==8.3.1
dj-database-url==3.0.1
Django>=5.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.1
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
packaging==25.0
pillow==12.0.0
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
safetensors==0.7.0
segmentation_models_pytorch==0.5.0
shellingham==1.5.4
sqlparse==0.5.4
sympy==1.14.0
timm==1.0.22
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
SimpleITK
pydicom
